@@ -1,71 +1,46 @@
 {
-  "name": "codex-website",
-  "description": "A really cool Vue.js project",
-  "version": "1.0.0",
-  "author": "Manuela <manu.cm43@gmail.com>",
-  "license": "MIT",
+  "name": "codexudea.github.io",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --open --hot",
-    "build": "cross-env NODE_ENV=production webpack --progress --hide-modules",
-    "lint:vue": "yarn eslint ./src/**/*.vue",
-    "contributors:add": "all-contributors add",
-    "contributors:generate": "all-contributors generate"
-  },
-  "engines": {
-    "node": ">=8.12"
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "all-contributors-cli": "^6.20.0",
-    "flexboxgrid": "^6.3.1",
-    "font-awesome": "^4.7.0",
-    "node": "8.12.0",
-    "normalize.css": "^8.0.0",
-    "vue": "^2.5.11",
-    "vue-router": "^3.0.1",
-    "yarn": "^1.15.2"
+    "core-js": "^3.6.5",
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0-0"
+  },
+  "devDependencies": {
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-router": "^4.5.13",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^6.7.2",
+    "eslint-plugin-vue": "^7.0.0",
+    "sass": "^1.41.1",
+    "sass-loader": "^10.2.0"
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/vue3-essential",
+      "eslint:recommended"
+    ],
+    "parserOptions": {
+      "parser": "babel-eslint"
+    },
+    "rules": {}
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not ie <= 8"
-  ],
-  "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-stage-3": "^6.24.1",
-    "cross-env": "^5.0.5",
-    "css-loader": "^0.28.7",
-    "eslint": "^5.7.0",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-friendly-formatter": "^4.0.1",
-    "eslint-loader": "^2.1.1",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-node": "^7.0.1",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^4.7.1",
-    "file-loader": "^1.1.4",
-    "husky": "^7.0.2",
-    "lint-staged": "7.3.0",
-    "node-sass": "^6.0.1",
-    "sass-loader": "^6.0.6",
-    "standard": "^12.0.1",
-    "vue-loader": "^13.0.5",
-    "vue-template-compiler": "^2.4.4",
-    "webpack": "^3.6.0",
-    "webpack-dev-server": "^2.9.1"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "*.vue": [
-      "eslint --fix",
-      "git add"
-    ]
-  }
+    "not dead"
+  ]
 }
