@@ -1,11 +1,3 @@
 <template>
-  <div id="app">
-    <router-view></router-view>
-  </div>
+  <router-view></router-view>
 </template>
-
-<script>
-export default {
-  name: 'app'
-}
-</script>
